Finished day 12 and changed i32 to usize for day IDs

+++ src/day/day_02.rs
@@ -13,7 +13,7 @@ lazy_static! {
 }
 
 impl AoCDay for Day02 {
-    fn day(&self) -> i32 {
+    fn day(&self) -> usize {
         2
     }
     fn part1(&self) -> String {
